refactor: 패키지 변경 및 네이밍 수정
